refactoring settings modal
